REFACTOR: Separate backend procedures from RPC interface.
